revert unnecessary changes to the test files
