+++ plugin/evm/gossip_test.go
@@ -5,9 +5,7 @@ package evm
 
 import (
 	"context"
-	"crypto/ecdsa"
 	"math/big"
-	"strings"
 	"testing"
 	"time"
 
@@ -28,29 +26,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/require"
 )
-
-func getValidEthTxs(key *ecdsa.PrivateKey, count int, gasPrice *big.Int) []*types.Transaction {
-	res := make([]*types.Transaction, count)
-
-	to := common.Address{}
-	amount := big.NewInt(0)
-	gasLimit := uint64(37000)
-
-	for i := 0; i < count; i++ {
-		tx, _ := types.SignTx(
-			types.NewTransaction(
-				uint64(i),
-				to,
-				amount,
-				gasLimit,
-				gasPrice,
-				[]byte(strings.Repeat("aaaaaaaaaa", 100))),
-			types.HomesteadSigner{}, key)
-		tx.SetFirstSeen(time.Now().Add(-1 * time.Minute))
-		res[i] = tx
-	}
-	return res
-}
 
 func TestGossipAtomicTxMarshaller(t *testing.T) {
 	require := require.New(t)
